chore(toolkit): Cleanup and improve codebase for Toolkit with Nextjs app router (#453)
* chore(toolkit): cleanup app router code from migration

* simplify chat routing
