+++ src/interfaces/coral_web/src/app/(auth)/layout.tsx
@@ -1,0 +1,7 @@
+import { WelcomePage } from '@/components/WelcomePage';
+
+const AuthLayout: React.FC<React.PropsWithChildren> = ({ children }) => {
+  return <WelcomePage>{children}</WelcomePage>;
+};
+
+export default AuthLayout;
