+++ src/interfaces/coral_web/src/app/(auth)/logout/page.tsx
@@ -1,36 +1,13 @@
-'use client';
+import { Metadata } from 'next';
 
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
+import Logout from './Logout';
 
-import { Text } from '@/components/Shared';
-import { WelcomePage } from '@/components/WelcomePage';
-import { useSession } from '@/hooks/session';
+export const metadata: Metadata = {
+  title: 'Logout',
+};
 
-/**
- * @description The login page supports logging in with an email and password.
- */
 const LogoutPage = () => {
-  const router = useRouter();
-  const { logoutMutation } = useSession();
-
-  useEffect(() => {
-    logoutMutation.mutate(undefined, {
-      onSettled: () => {
-        router.push('/login');
-      },
-    });
-  }, []);
-
-  return (
-    <WelcomePage title="Logout" navigationAction="login">
-      <div className="flex flex-col items-center justify-center">
-        <Text as="h1" styleAs="h3">
-          Logging out
-        </Text>
-      </div>
-    </WelcomePage>
-  );
+  return <Logout />;
 };
 
 export default LogoutPage;
