+++ src/interfaces/coral_web/src/app/(main)/discover/page.tsx
@@ -1,4 +1,10 @@
+import { Metadata } from 'next';
+
 import { DiscoverAgents } from '@/components/Agents/DiscoverAgents';
+
+export const metadata: Metadata = {
+  title: 'Discover Assistants',
+};
 
 const DiscoverAssistantPage: React.FC = () => {
   return <DiscoverAgents />;
