+++ src/interfaces/coral_web/src/app/(main)/new/page.tsx
@@ -1,4 +1,10 @@
+import { Metadata } from 'next';
+
 import { CreateAgent } from '@/components/Agents/CreateAgent';
+
+export const metadata: Metadata = {
+  title: 'New Assistant',
+};
 
 const NewAssistantPage: React.FC = () => {
   return <CreateAgent />;
