+++ src/interfaces/coral_web/src/components/Layout.tsx
@@ -2,7 +2,6 @@
 'use client';
 
 import { Transition } from '@headlessui/react';
-import { capitalize } from 'lodash';
 import React, { Children, PropsWithChildren, useContext, useEffect, useState } from 'react';
 
 import { AgentsSidePanel } from '@/components/Agents/AgentsSidePanel';
@@ -13,7 +12,6 @@ import { NavigationUserMenu } from '@/components/NavigationUserMenu';
 import { SettingsDrawer } from '@/components/Settings/SettingsDrawer';
 import { Banner } from '@/components/Shared';
 import { NavigationBar } from '@/components/Shared/NavigationBar/NavigationBar';
-import { PageHead } from '@/components/Shared/PageHead';
 import { BannerContext } from '@/context/BannerContext';
 import { useIsDesktop } from '@/hooks/breakpoint';
 import { useSession } from '@/hooks/session';
@@ -23,16 +21,14 @@ import { cn } from '@/utils/cn';
 export const LeftSection: React.FC<React.PropsWithChildren> = ({ children }) => <>{children}</>;
 export const MainSection: React.FC<React.PropsWithChildren> = ({ children }) => <>{children}</>;
 
-type LayoutProps = {
-  title?: string;
-} & PropsWithChildren;
+type LayoutProps = {} & PropsWithChildren;
 
 /**
  * This component is in charge of layout out the entire page.
  * It shows the navigation bar, the left drawer and main content.
  * On small devices (e.g. mobile), the left drawer and main section are stacked vertically.
  */
-export const Layout: React.FC<LayoutProps> = ({ title = 'Chat', children }) => {
+export const Layout: React.FC<LayoutProps> = ({ children }) => {
   const { message: bannerMessage } = useContext(BannerContext);
   const {
     settings: { isConvListPanelOpen, isMobileConvListPanelOpen },
@@ -69,7 +65,6 @@ export const Layout: React.FC<LayoutProps> = ({ title = 'Chat', children }) => {
 
   return (
     <>
-      <PageHead title={capitalize(title)} />
       <div className="flex h-screen w-full flex-1 flex-col gap-3 bg-mushroom-900 p-3">
         <NavigationBar>
           <span className="flex items-center gap-x-2">
@@ -179,7 +174,6 @@ export const AgentsLayout: React.FC<AgentsLayoutProps> = ({
 
   return (
     <>
-      <PageHead title={capitalize(title)} />
       <div className="dark:bg-vb-60 flex h-screen w-full flex-1 flex-col gap-3 bg-mushroom-900 p-3">
         <div
           className={cn(
